Enhance documentation for safe addition and subtraction methods
- Added detailed explanations of use cases for safeAddition and safeSubtraction methods in UInt64 extension.
- Clarified the behavior of these methods when handling overflow and underflow.
- Highlighted their importance in contexts requiring safe and predictable arithmetic operations.
